requirements.txt
pyyaml
requests
django-cors-headers